fixed bug with producer not delivering energy if radiation is < 500 and removed unneccessary codelines

+++ energy-community-producer/src/main/java/com/example/energy_community_producer/EnergyProductionScheduler.java
@@ -19,7 +19,7 @@ public class EnergyProductionScheduler {
     @Scheduled(fixedDelay = 3000)
     public void produce() {
 
-        int shortwaveRadiationFactor = this.shortwaveRadiationService.getShortwaveRadiationFactor();
+        double shortwaveRadiationFactor = this.shortwaveRadiationService.getShortwaveRadiationFactor();
 
         double random = Math.random() * 2;
